Adding XML well-formedness parsing for XML docs with ideas from Jacques Distler http://golem.ph.utexas.edu/~distler/blog/archives/001054.html
We could be using only the XML parser and not parse twice, but
at the moment I am tempted to keep using opensp, for as long
as we have the best library of error message and explanations for it.

Code is not really pretty, and should be amended at some point to
use a real SAX ErrorHandler. That said, it very much tolls the bell for the
"Validator XML support has some limitations." message.